Intermediate commit: m64 working but not good
m64 testing is working for e10 > 0, but it's clear that this route is
not really sustainable without building a ton of branches into the
testing, which makes it hard to tell if the test is actually checking
the right thing.

Use this as a save point to try to get tests working using BigNum-based
rational numbers, which should make the tests much more expressive.